Move Output to separate namespace
